添加回复的事务逻辑 添加异常处理 修改数据库中id的类型 改为Long

+++ src/main/java/com/community/model/User.java
@@ -4,12 +4,12 @@ import lombok.Data;
 
 @Data
 public class User {
-    private Integer id;
+    private Long id;
     private String name;
     private String accountId;
     private String token;
-    private long gmtCreate;
-    private long gmtModified;
+    private Long gmtCreate;
+    private Long gmtModified;
     private String avatarUrl;
 
 }
